Replace smbase custom string with std::string.

+++ src/smbase/bitarray.cc
@@ -158,7 +158,7 @@ string toString(BitArray const &b)
   int len = b.length();
   stringBuilder ret(len);
   for (int i=0; i<len; i++) {
-    ret[i] = b.test(i)? '1' : '0';
+    ret[size_t(i)] = b.test(i)? '1' : '0';
   }
   return ret;
 }
